requirements.txt
groq>=0.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
flask>=3.0.3
requests>=2.31.0
pydantic>=2.7.0
python-dotenv>=1.0.1
reportlab>=4.1.0
jinja2>=3.1.4
colorama>=0.4.6
tabulate>=0.9.0
